requirements.txt
# requirements.txt
requests
sentence-transformers
faiss-cpu
numpy
fastapi
uvicorn
jinja2
httpx
python-dotenv